Install Ruby 1.9.3-p545 on Darwin when rvm = 2.0.0
When the Mac VM was updated, rvm was also updated.
On this new version, 1.9.3 points to p545, which is not
installed on the Mac VM.

This leads to a build error seen at
https://travis-ci.org/travis-ci/travis.rb/jobs/23385233

The idea here is to install the desirable version of Ruby on the Mac
if we are running with Ruby 2.0.0.